Add some symlinks to preserve backwards compatability.
