--- bmrb
@@ -1,1 +1,0 @@
-pynmrstar/
